graph, graphql: Use Arc<ObjctType> to reference object types
Rather than use a string name, use the actual object type to identify
types. It's not possible to do this with plain references, for example,
because we pass a reference to a SelectionSet to graph::spawn_blocking, so
we do the next best thing and use an Arc. Unfortunately, because
`graphql_parser` doesn't wrap its object types in an Arc, that means we
need to keep a copy of all of them in ApiSchema.